# api: Do more caching to speed up search responses