--- a/libs/api.lua
+++ b/libs/api.lua
@@ -83,6 +83,8 @@ local installDeps = require('install-deps').toDb
 local ffi = require('ffi')
 local fs = require('coro-fs')
 local metrics = require('metrics')
+local cachedDb = require('db-cached')
+local uv = require('uv')
 
 local function hex_to_char(x)
   return string.char(tonumber(x, 16))
@@ -129,6 +131,7 @@ local function compileGlob(glob)
 end
 
 local metaCache = {}
+local urlCache = {}
 
 -- Define the required metrics that must exist for collectMetrics() to work.
 metrics.define("lua.mem.used")
@@ -189,8 +192,12 @@ end
 
 return function (db, prefix)
 
+  local function makeUrlPath(kind, hash, filename)
+    return "/" .. kind .. "s/" .. hash .. '/' .. filename
+  end
+
   local function makeUrl(kind, hash, filename)
-    return prefix .. "/" .. kind .. "s/" .. hash .. '/' .. filename
+    return prefix .. makeUrlPath(kind, hash, filename)
   end
 
   local function loadMeta(author, name, version)
@@ -219,10 +226,26 @@ return function (db, prefix)
     if meta.type == "blob" then
       filename = filename .. ".lua"
     end
-    meta.url = makeUrl(meta.type, meta.object, filename)
+    -- because prefix can change per-request, cache the path
+    -- separately and prefix it at response-time
+    local urlPath = makeUrlPath(meta.type, meta.object, filename)
+    urlCache[meta] = urlPath
+    meta.url = urlPath
     metaCache[hash] = meta
     return meta
   end
+
+  -- use cached db
+  db = cachedDb(db)
+
+  local timeStart, memStart = uv.now(), collectgarbage("count")
+  -- warm up db cache and meta cache
+  for author in db.authors() do
+    for name in db.names(author) do
+      local _ = loadMeta(author, name)
+    end
+  end
+  print('cache warmed in ' .. (uv.now() - timeStart) .. 'ms, using ' .. string.format("%0.2f", (collectgarbage("count") - memStart) / 1024) .. ' MB memory')
 
   local routes = {
     "^/metrics$", collectStats,
@@ -500,6 +523,10 @@ return function (db, prefix)
       end
     end
     if type(body) == "table" then
+      -- update all urls to use the current prefix
+      for match, meta in pairs(body.matches) do
+        meta.url = (prefix or '') .. urlCache[meta]
+      end
       body = jsonStringify(body) .. "\n"
       res.headers["Content-Type"] = "application/json"
     end