Should help with https://github.com/luvit/luvit.io/issues/24

There are two parts to this:

- The API was already caching the `hash -> meta` part of the `author/name/version -> hash -> meta` lookup, but this cache was not being warmed on start, so the first request would take roughly twice as long as future requests.
  + Warming the cache in advance also required a change to how urls were cached, since its technically possible for `prefix` to change between requests, so we can't use the same cached url every time
- All db lookups would access the file system every time, meaning that all `/search/` endpoint requests would iterate all refs, read them, etc for general searches i.e. `/search/*`. The actual values very rarely change, though, and the amount of memory needed to cache all those lookups is rather small (~1.5MB for lit-backup.git), so we can cache them and just invalidate the cache when a new ref is added via `db.write`.
  + This cache is also warmed at the same time as the `hash -> meta` one

Performance numbers for `/search/*`:

Before:
```
request 1 handled in 402ms, body length = 182518
request 2 handled in 209ms, body length = 182518
request 3 handled in 216ms, body length = 182518
request 4 handled in 206ms, body length = 182518
request 5 handled in 205ms, body length = 182518
```

After:
```
cache warmed in 444ms, using 1.52 MB memory
request 1 handled in 22ms, body length = 182518
request 2 handled in 21ms, body length = 182518
request 3 handled in 18ms, body length = 182518
request 4 handled in 19ms, body length = 182518
request 5 handled in 18ms, body length = 182518
```

(almost all the time for `/search/*` is now spent in `json.stringify`)

---

This is currently a draft because I haven't tested the cache invalidation part of it yet (`db.write`), and am not positive that `db.write` is the only pathway to modifying the cached data (if other `db` functions could modify the cached data, then the cache would not be updated, leading to the lit api responses being out-of-sync with the actual state of the db).