@squeek502 2020-02-28
Code used for benchmarking (put in `bench.lua` in the `commands` dir and run via `luvi . -- bench`):

```lua
local args = {...}
return function ()
  local uv = require 'uv'
  local core = require('core')()
  local db = core.db
  local handleRequest = require('api')(db, '')

  local req = {
    method = "GET",
    path = "/search/*",
    headers = {Host = "localhost"},
  }

  local lastBody = nil
  for i=1,5 do
    local res = {
      headers = {},
    }

    local start = uv.now(uv.update_time())
    handleRequest(req, res)
    local time = uv.now(uv.update_time()) - start
    print('request '..i..' handled in ' .. time .. 'ms, body length = ' .. #res.body)
    assert(not lastBody or res.body == lastBody, 'response body mismatch between req ' .. (i-1) .. ' and req ' .. i)
    lastBody = res.body
  end
end
```