@squeek502 2020-03-02
Okay, I confirmed that cache invalidation works when a package is published. Tested via:

- copied `litdb.git` to `litdb-server.git`
- made a separate `litconfig-server` with the `database` key set to the path to `litdb-server.git`
- set env var `LIT_CONFIG` to the path to `litconfig-server` and did `luvi . -- serve` (so the server uses the `litdb-server.git` repo)
- unset the `LIT_CONFIG` var
- `luvi . -- up ws://localhost:4822` to set the upstream in the default litconfig to the local one we just setup via `serve`
- Add/publish a new package via `luvi . -- publish test.lua` to the local upstream

With some debug prints added in `db.write` in `db-cached.lua`, the server outputs:

```
client command: 127.0.0.1:40849 - send
'db.write'      'squeek502'     'test'  '0.0.6' '27c68bcf34ce2619f3469c55f7ce41c1e712d46e'
'db.write versions cache before:'       { '0.0.1', '0.0.2', '0.0.3', '0.0.4', '0.0.5' }
'db.write versions cache after:'        { '0.0.1', '0.0.2', '0.0.3', '0.0.4', '0.0.5', '0.0.6' }
new package: squeek502/test/v0.0.6
client disconnected: 127.0.0.1:40849
```

and the new package shows up everywhere you'd expect (`/search/name:test`, `/packages/squeek502`, etc).

This is ready for review.